move empty_letters out of adobefonts directory
